refactor(TemplateBuilder): convert buildPipeElements to PipeList component and extract usePipeTypeSearch hook
  - Replace buildPipeElements function with PipeList React.FC; move useStyles2 inside and drop useMemo wrapper in TemplateQueryEditor
  - Extract search/filter logic (keyword groups, label/description match, navItems, flatIndexMap) from PipeTypeSearchMenu into new usePipeTypeSearch hook
  - Memoize displayGroups properly (was recomputed every render via IIFE)
  - PipeTypeSearchMenu is now a pure UI component: state, focus, keyboard wiring, and render
